fixed chai import error
