+++ back-end/package.json
@@ -6,11 +6,12 @@
   "type": "module",
   "scripts": {
     "watch": "nodemon ./server.js",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "mocha --exit",
     "start": "node server.js",
-    "dev": "nodemon server.js"
+    "dev": "nodemon server.js",
+    "coverage": "c8 npm test"
   },
-  "author": "",
+  "author": "c",
   "license": "ISC",
   "dependencies": {
     "axios": "^1.8.4",
@@ -23,7 +24,9 @@
   },
   "devDependencies": {
     "chai": "^5.2.0",
+    "chai-http": "^4.4.0",
     "mocha": "^11.1.0",
-    "nodemon": "^3.1.9"
+    "nodemon": "^3.1.9",
+    "sinon": "^20.0.0"
   }
 }
