+++ back-end/routes/friend/getUser.js
@@ -1,3 +1,4 @@
+//getUser.js
 import express from 'express';
 import axios from 'axios';
 import dotenv from 'dotenv';
