+++ back-end/test/test.js
@@ -1,6 +1,8 @@
+/*
 import { use, expect, should, assert } from 'chai'
 import { default as chaiHttp, request } from 'chai-http'
 import { parse } from 'node-html-parser'
 import app from '../app.js'
 
 use(chaiHttp)
+*/
